pyproject.toml
[tool.uv]
package = false

[project]
name = "hatchup-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "alireza",email = "alirezaghanati@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"


dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (==3.16.0)",
    "drf-spectacular (==0.27.2)",
    "djangorestframework-simplejwt (==5.5.1)",
    "django-redis (==5.4.0)",
    "django-fsm-2 (==4.1.0)",
    "psycopg2-binary (==2.9.11)",
    "python-dotenv (==1.0.1)",
    "pillow (==11.1.0)",
    "django-storages (==1.14.2)",
    "boto3 (==1.34.59)",
    "django-jsonform (==2.23.2)",
    "django-money (==3.5.4)",
    "django-filter (==25.2)",
    "drf-yasg (==1.21.14)",
    "python-magic (==0.4.27)",
    "django-ckeditor-5 (==0.2.19)",
    "six (==1.17.0)",
    "sentry-sdk (>=2.0.0,<3.0.0)",
    "django-cors-headers (>=4.3.0)",
    "whitenoise (>=6.8.0)",
    "argon2-cffi (>=23.1.0)",
]




